server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.15.0
stripe>=12.0.0
resend>=2.0.0
pytest==9.0.2
pytest-cov==7.1.0
pytest-asyncio==1.3.0
httpx==0.28.1
